+++ src/InstanceReader.java
@@ -18,12 +18,12 @@ public class InstanceReader {
             k = scan.nextInt();
 
             grafo = new Graph(numVertices);
-            
+
             for(int i = 0; i < numArestas; i++){
                 // Vai ler todas as arestas u -> v e guardar seus pesos
                 int u = scan.nextInt();
                 int v = scan.nextInt();
-                int peso = scan.nextInt();
+                double peso = scan.nextDouble();
                 
                 grafo.addDirectEdge(u, v, peso);
             }
@@ -32,7 +32,7 @@ public class InstanceReader {
         //Transforma o grafo em uma matriz de distâncias mínimas
         grafo.runFloydWarshall();
 
-        int[][] matriz = grafo.getMatrizDisntancia();
+        double[][] matriz = grafo.getMatrizDisntancia();
 
         return new InstanceData(numVertices, k, matriz);
     }
